Set up basic MVC structure for HouseRentingSystem
This commit includes the creation of necessary Controllers, Views, and Models for the House and Agent modules. Also, adjustments have been made to existing components such as HomeController and MainMenuComponent to reflect the updated menu options and dependencies. Noteworthy, changes to SignIn options in the RentingSystemServiceCollectionExtension.cs have also been implemented, simplifying the password requirements for ease of development and testing.