Implement house service and update home page view
A new house renting service is implemented along with related contract and models. The HomeController now uses this House Service to display the last three houses on the home page view, which has also been updated to include a carousel display of houses. In addition, the house renting service is registered in the ServiceCollectionExtension.