Implement Agent service and update related components
The AgentService class has been implemented with essential methods such as ExistsByIdAsync and CreateAsync. This was done in connection with the IRepository. Additionally, a unique index enforcement has been added to the Agent model for the PhoneNumber property. Methods in HouseService and IHouseService have been renamed to follow an asynchronous naming convention, and logic to check if a user already exists as an agent has been added in AgentController. A new ClaimsPrincipalExtensions class has also been created for reusability purposes.